# src/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This page is API documentation for a little game like predict age, gender by name, get random fact, image, joke",
        "title": "GOGAME",
        "contact": {
            "name": "Developer",
            "email": "intanmarsjaf@outlook.com"
        },
        "version": "1.0"
    },
    "host": "localhost:1111",
    "paths": {
        "/game/dare": {
            "get": {
                "description": "This is API to get random activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Random Activity",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/game/fact": {
            "get": {
                "description": "This is API to get random fact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Random Fact",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/game/img": {
            "get": {
                "description": "This is API to get random image",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Random Image",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/game/joke": {
            "post": {
                "description": "This is API to get random joke",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Random Joke",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "domain.InputRandomJoke",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.InputRandomJoke"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/game/predict": {
            "post": {
                "description": "This is API to get prediction from name input by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Prediction By Name",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "domain.ExamplePredictName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ExamplePredictName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/master/api": {
            "get": {
                "description": "This is API to get list of public API by auth or category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "List Public API",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Params 1",
                        "name": "auth",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Params 2",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/master/country": {
            "get": {
                "description": "This is API to get list country by country id or region",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Data"
                ],
                "summary": "List Country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Params 1",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Params 2",
                        "name": "region",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/twitter/timeline": {
            "post": {
                "description": "This is API to see tweets in timeline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twitter"
                ],
                "summary": "Get Tweets in Timeline",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "domain.ExamplePredictName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.InputTimeline"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/twitter/tweet/fact": {
            "get": {
                "description": "This is API to tweet random fact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Twitter"
                ],
                "summary": "Random Fact",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadRequestResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.BadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "004"
                },
                "message": {
                    "type": "string",
                    "example": "[message]"
                },
                "result": {
                    "type": "string",
                    "example": "null"
                }
            }
        },
        "domain.ExamplePredictName": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Admin"
                }
            }
        },
        "domain.InputRandomJoke": {
            "type": "object",
            "properties": {
                "is_safe": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "domain.InputTimeline": {
            "type": "object",
            "properties": {
                "total_data_show": {
                    "type": "integer"
                }
            }
        },
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "000"
                },
                "message": {
                    "type": "string",
                    "example": "[message]"
                },
                "result": {
                    "type": "object"
                }
            }
        }
    }
}